no need for getHeight() to return a long

+++ core/api/src/main/java/hu/netmind/bitcoin/BlockChain.java
@@ -119,6 +119,6 @@ public interface BlockChain extends Observable
     * Returns the height of the best chain
     * @return 
     */
-   long getHeight();
+   int getHeight();
    
 }
